refactor: fix file naming, update tests
