+++ web/src/components/task/task-sidebar-flow/__tests__/annotation-list.test.tsx
@@ -32,6 +32,7 @@ describe('AnnotationList', () => {
         onLinkDeleted: () => {},
         onLabelSelect: () => {},
         onLabelDelete: () => {},
+        onAnnotationDeleted: () => {},
         isOwner: false,
         labels: []
     };
